+++ engine/renderer/vulkan/swapchain.h
@@ -59,7 +59,7 @@ namespace egkr
 		egkr::vector<texture*> render_textures_{};
 		egkr::vector<texture*> depth_attachments_{};
 
-		uint32_t image_count_{};
+		uint8_t image_count_{};
 		uint8_t max_frames_in_flight_{3};
 
 	};
