+++ engine/systems/view_system.cpp
@@ -72,7 +72,7 @@ namespace egkr
 		}
 
 		view_system_->registered_views_.push_back(render_view);
-		view_system_->registered_views_by_name_[configuration.name] = view_system_->registered_views_.size() - 1;
+		view_system_->registered_views_by_name_[configuration.name] = (uint32_t)view_system_->registered_views_.size() - 1;
 
 	}
